remove unused code reorga types pkg
