+++ frontend/src/components/messstelle/charts/MessstelleDiagramme.vue
@@ -134,13 +134,13 @@ import LadeZaehldatumDTO from "@/types/zaehlung/zaehldaten/LadeZaehldatumDTO";
 import MesswerteListenausgabe from "@/components/messstelle/charts/MesswerteListenausgabe.vue";
 import BelastungsplanMessquerschnitteDTO from "@/types/messstelle/BelastungsplanMessquerschnitteDTO";
 import BelastungsplanMessquerschnittCard from "@/components/messstelle/charts/BelastungsplanMessquerschnittCard.vue";
-import MessstelleHistoryItem from "@/types/app/MessstelleHistoryItem";
+import MessstelleHistoryItem from "@/types/history/MessstelleHistoryItem";
 import MessstelleInfoDTO from "@/types/messstelle/MessstelleInfoDTO";
 import MessstelleOptionsDTO from "@/types/messstelle/MessstelleOptionsDTO";
 import _ from "lodash";
 import PdfReportMenueMessstelle from "@/components/messstelle/PdfReportMenueMessstelle.vue";
 import GenerateCsvService from "@/api/service/GenerateCsvService";
-import CsvDTO from "@/types/CsvDTO";
+import CsvDTO from "@/types/common/CsvDTO";
 import BannerMesstelleTabs from "@/components/messstelle/charts/BannerMesstelleTabs.vue";
 import GeneratePdfService from "@/api/service/GeneratePdfService";
 import DaveUtils from "@/util/DaveUtils";
