+++ frontend/src/components/messstelle/optionsmenue/panels/ZeitPanel.vue
@@ -92,7 +92,7 @@
 import PanelHeader from "@/components/common/PanelHeader.vue";
 import { computed, onMounted, ref, Ref, watch } from "vue";
 import MessstelleOptionsmenuService from "@/api/service/MessstelleOptionsmenuService";
-import NichtPlausibleTageDTO from "@/types/NichtPlausibleTageDTO";
+import NichtPlausibleTageDTO from "@/types/messstelle/NichtPlausibleTageDTO";
 import { useStore } from "@/api/util/useStore";
 import MessstelleOptionsDTO from "@/types/messstelle/MessstelleOptionsDTO";
 import { useDateUtils } from "@/util/DateUtils";
