+++ frontend/src/components/zaehlstelle/optionsmenue/panels/ZeitauswahlPanel.vue
@@ -174,7 +174,7 @@ import ZaehldatenIntervall, {
     ZaehldatenIntervallToSelect,
 } from "@/types/enum/ZaehldatenIntervall";
 import Zeitblock, { zeitblockInfo } from "@/types/enum/Zeitblock";
-import KeyVal from "@/types/KeyVal";
+import KeyVal from "@/types/common/KeyVal";
 import ZeitblockStuendlich, {
     zeitblockStuendlichInfo,
 } from "@/types/enum/ZeitblockStuendlich";
