+++ frontend/src/types/app/SsoUserInfoResponse.ts
