+++ frontend/src/types/app/VersionInfoResponse.ts
