+++ frontend/src/types/common/BaseEntity.ts
