+++ frontend/src/types/enum/AuswertungCategories.ts
@@ -1,5 +1,3 @@
-import Himmelsrichtungen from "@/types/enum/Himmelsrichtungen";
-
 export enum Monate {
     JANUAR = "Januar",
     FEBRUAR = "Februar",
@@ -33,24 +31,3 @@ export enum ZeitintervallCategories {
     QUARTALE = "Quartale",
     MONATE = "Monate",
 }
-
-export const categorieAsNumber: Map<string, number> = new Map<string, number>([
-    [Monate.JANUAR, 0],
-    [Monate.FEBRUAR, 1],
-    [Monate.MAERZ, 2],
-    [Monate.APRIL, 3],
-    [Monate.MAI, 4],
-    [Monate.JUNI, 5],
-    [Monate.JULI, 6],
-    [Monate.AUGUST, 7],
-    [Monate.SEPTEMBER, 8],
-    [Monate.OKTOBER, 9],
-    [Monate.NOVEMBER, 10],
-    [Monate.DEZEMBER, 11],
-    [Quartale.QUARTAL_1, 2],
-    [Quartale.QUARTAL_1, 2],
-    [Quartale.QUARTAL_1, 2],
-    [Quartale.QUARTAL_1, 2],
-    [Halbjahre.HALBJAHR_1, 5],
-    [Halbjahre.HALBJAHR_2, 11],
-]);
