--- frontend/src/types/enum/Status.ts
@@ -1,7 +1,0 @@
-enum Status {
-    IMPORTED = "IMPORTED",
-    ACTIVE = "ACTIVE",
-    INACTIVE = "INACTIVE",
-}
-
-export default Status;
