+++ frontend/src/types/history/AbstractHistoryItem.ts
