+++ frontend/src/types/karte/AnzeigeKarteDTO.ts
