+++ frontend/src/types/karte/MessstelleKarteDTO.ts
@@ -1,5 +1,5 @@
-import TooltipMessstelleDTO from "@/types/TooltipMessstelleDTO";
-import AnzeigeKarteDTO from "@/types/AnzeigeKarteDTO";
+import TooltipMessstelleDTO from "@/types/karte/TooltipMessstelleDTO";
+import AnzeigeKarteDTO from "@/types/karte/AnzeigeKarteDTO";
 import { MessstelleStatus } from "@/types/enum/MessstelleStatus";
 
 export default interface MessstelleKarteDTO extends AnzeigeKarteDTO {
