+++ frontend/src/types/karte/TooltipMessstelleDTO.ts
