+++ frontend/src/types/karte/TooltipZaehlstelleDTO.ts
