+++ frontend/src/types/karte/ZaehlstelleKarteDTO.ts
@@ -1,6 +1,6 @@
 import ZaehlartenKarteDTO from "@/types/zaehlstelle/ZaehlartenKarteDTO";
-import TooltipZaehlstelleDTO from "@/types/TooltipZaehlstelleDTO";
-import AnzeigeKarteDTO from "@/types/AnzeigeKarteDTO";
+import TooltipZaehlstelleDTO from "@/types/karte/TooltipZaehlstelleDTO";
+import AnzeigeKarteDTO from "@/types/karte/AnzeigeKarteDTO";
 
 export default interface ZaehlstelleKarteDTO extends AnzeigeKarteDTO {
     letzteZaehlungId: string;
