+++ frontend/src/types/suche/Suggest.ts
