--- frontend/src/types/zaehlstelle/GeoPoint.ts
@@ -1,4 +1,0 @@
-export default interface GeoPoint {
-    lat: string;
-    lon: string;
-}
