+++ frontend/src/types/zaehlung/LadeFahrbeziehungDTO.ts
@@ -1,4 +1,4 @@
-import BaseEntity from "@/types/BaseEntity";
+import BaseEntity from "@/types/common/BaseEntity";
 
 export default interface LadeFahrbeziehungDTO extends BaseEntity {
     // Kreuzung
